# src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: CITSA Mobile Backend API
  description: |
    REST API for the CITSA (Computer and Information Technology Students Association) Mobile Application.

    ## Authentication
    The API uses **JWT Bearer Token** authentication via a 2-step OTP flow:
    1. Call `POST /auth/send-otp` with your student ID to receive a 6-digit OTP via email
    2. Call `POST /auth/verify-otp` with the student ID and OTP code to receive access + refresh tokens
    3. Include the access token in the `Authorization: Bearer <token>` header for protected endpoints
    4. When the access token expires (1h), use `POST /auth/refresh-token` to get a new one

    ## Student ID Format
    Must match pattern: `XX/XXX/YY/NNNN` (e.g., `PS/ITC/22/0120`)

    ## Rate Limiting
    - General: 100 requests per 15 minutes per IP
    - OTP: 3 requests per 5 minutes per IP
    - Auth: 5 requests per 15 minutes per IP
    - Uploads: 10 requests per 15 minutes per IP

    ## Pagination
    Paginated endpoints accept `page` (default: 1) and `limit` (default: 20, max: 100) query params.
    Responses include a `pagination` object with `page`, `limit`, `total`, `totalPages`, `hasNext`, `hasPrev`.
  version: 1.0.0
  contact:
    name: CITSA Dev Team
    email: biileprinceyennuyar5@gmail.com
  license:
    name: ISC
servers:
  - url: https://citsa-mobile-backend.onrender.com/api/v1
    description: Production Server (Render)
  - url: http://localhost:3000/api/v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from verify-otp or refresh-token
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false

    # ─── Auth ──────────────────────────────────────
...
  responses:
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
...
paths:
...
  /groups:
    get:
      tags: [Groups]
      summary: Get all groups
      description: |
        Returns paginated groups with optional filtering.
        Authentication is optional — if provided, `isMember` and `userRole` will be populated.
        Ordered by member count descending, then name ascending.
      operationId: getGroups
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by group category
        - name: search
          in: query
          schema:
            type: string
          description: Search in name and description
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Group"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
    post:
      tags: [Admin]
      summary: Create group (Admin only)
      description: Create a new group. The admin is automatically added as the group admin member.
      operationId: adminCreateGroup
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Group created successfully"
                  data:
                    $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /groups/my-groups:
    get:
      tags: [Groups]
      summary: Get my groups
      description: Returns groups that the authenticated user is a member of.
      operationId: getMyGroups
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: User's groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/Group"
                        - type: object
                          properties:
                            isMember:
                              type: boolean
                              example: true
                            userRole:
                              type: string
                              enum: [MEMBER, MODERATOR, ADMIN]
                            joinedAt:
                              type: string
                              format: date-time
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /groups/{id}:
    get:
      tags: [Groups]
      summary: Get single group
      description: Returns group details. Authentication is optional.
      operationId: getGroupById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/Group"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Admin]
      summary: Update group (Admin only)
      description: Update group details. Checks for duplicate name on rename.
      operationId: adminUpdateGroup
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupRequest"
      responses:
        "200":
          description: Group updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Group updated successfully"
                  data:
                    $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Admin]
      summary: Delete group (Admin only)
      description: Permanently delete a group and all associated memberships.
      operationId: adminDeleteGroup
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Group deleted successfully"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /groups/{id}/members:
    get:
      tags: [Groups]
      summary: Get group members
      description: Returns a paginated list of group members ordered by role (admins first) then join date.
      operationId: getGroupMembers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Group members list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/GroupMember"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "404":
          $ref: "#/components/responses/NotFound"
  /groups/{id}/messages:
    get:
      tags: [Groups]
      summary: Get group messages
      description: Returns paginated group messages. Only group members can access messages.
      operationId: getGroupMessages
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Group messages list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/GroupMessage"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Groups]
      summary: Post group message (group admin only)
      description: Creates a group message. Only users with group role ADMIN can post.
      operationId: createGroupMessage
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupMessageRequest"
      responses:
        "201":
          description: Group message created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Message posted successfully"
                  data:
                    $ref: "#/components/schemas/GroupMessage"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /groups/{id}/messages/{messageId}/reactions:
    post:
      tags: [Groups]
      summary: React to group message
      description: Adds a like reaction to a group message. Only group members can react.
      operationId: reactToGroupMessage
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Reaction added
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Reaction added successfully"
                  data:
                    type: object
                    properties:
                      reacted:
                        type: boolean
                        example: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Groups]
      summary: Remove reaction from group message
      description: Removes the authenticated user's reaction from a group message.
      operationId: unreactToGroupMessage
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Reaction removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Reaction removed successfully"
                  data:
                    type: object
                    properties:
                      reacted:
                        type: boolean
                        example: false
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...